backend/src/csv_handler/requirements.txt
annotated-types==0.6.0
anyio==4.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
faker==23.2.1
fastapi==0.109.2
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
mangum==0.17.0
markdown-it-py==3.0.0
mdurl==0.1.2
psycopg2-binary==2.9.9
pydantic-core==2.16.2
pydantic==2.6.1
pygments==2.17.2
python-dateutil==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.6
requests==2.31.0
rich==13.7.0
six==1.16.0
sniffio==1.3.0
sqlalchemy==2.0.27
starlette==0.36.3
typing-extensions==4.9.0
urllib3==2.0.7
uvicorn==0.27.0.post1